Fixed file saving logic, added seconds to file names

+++ SystemInfo/export/screenCapture.cpp
@@ -7,7 +7,7 @@
 #include "../dialog/scrUploadDialog.h"
 #include "../network/rest/rest.h"
 
-void takeScreenshot(HWND hwnd, SCR_SAVETYPE scrSaveType) {
+ACTION takeScreenshot(HWND hwnd, SCR_SAVETYPE scrSaveType, RESULT_STRUCT *res) {
 	RECT winSize;
 	GetClientRect(hwnd, &winSize);
 	INT32 areaWidth = winSize.right - winSize.left;
@@ -26,29 +26,36 @@ void takeScreenshot(HWND hwnd, SCR_SAVETYPE scrSaveType) {
 	TCHAR fullSavePath[256];
 
 	ZeroMemory(&fullSavePath, sizeof(fullSavePath));
-	//to do delete the temp screenshot
 	if (scrSaveType == SCR_SAVETYPE::INTERNET) {
 		_tcscpy(fullSavePath, sysInfoConfigDirectoryPath);
 		TCHAR uuidString[256];
 		_tcscat(fullSavePath, L"\\");
 		_tcscat(fullSavePath, L"_temp.png");
 	} else if (scrSaveType == SCR_SAVETYPE::LOCAL) {
-		openFileDiag(hwnd, FILE_EXTENSION::PNG, fullSavePath, 1);
+		if (openFileDiag(hwnd, FILE_EXTENSION::PNG, fullSavePath, 1) == ACTION::CANCELED_OUT) {
+			return ACTION::CANCELED_OUT;
+		}
 	}
 
-	screen.Save(fullSavePath);
+	if (!SUCCEEDED(screen.Save(fullSavePath))) {
+		return ACTION::__ERROR;
+	}
 	DeleteObject(hCaptureBitmap);
 	ReleaseDC(hwnd, mainWindowDC);
 	DeleteDC(screenCapture);
-
+	if (scrSaveType == SCR_SAVETYPE::LOCAL) {
+		res->src.assign(fullSavePath);
+	}
 	if (scrSaveType == SCR_SAVETYPE::INTERNET) {
 		RESULT_STRUCT resStruct = {};
 		if (!uploadImage(&resStruct, fullSavePath)) {
 			MessageBox(NULL, _T("Failed to upload"), NULL, MB_ICONERROR | MB_OK);
+			return ACTION::__ERROR;
 		}
 		else {
 			_tcscpy(uploadSrcLinkData.link, resStruct.__src);
 		}
 		DeleteFile(fullSavePath);
 	}
+	return ACTION::ACCEPTED;
 }
